requirements.txt
langchain-google
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
sentence-transformers